Delete redundant per player CheatsEnable property
We do not see benefit in enabling cheats for certain players but not for the rest.
The AI cheats irrespective of the CheatsEnable property.

This patch uses InitAttributes to check if cheats are enabled making it
slightly harder to workaround it from js.

See SetLosRevealAll in CCmpAIManager and PETRA.Config.prototype.Cheat.

refs 996a32125ce7 / #879 (AI reveal los);
refs 202a02568b5 /  #1555 (cheat codes, cheatsEnable for AI);
refs 25293ce0cc5c  (AIDiff GatherRateMultiplier),
refs f6e0c4c5a408  / #2019 (no more cheatsEnable for AI),
refs 3246b02cdaec  (SetTradeRateMultiplier)
refs de72510c608b  / #6302 (move from InitGame.js to petra/config.js)