Never change the names in the "Player Name" column
The intent of this branch was to not change the names when a saved game
is loaded. So that one can see who was where in the saved game.
It was implemented wrong: The names didn't change when no savegame was
loaded and changed when a saved game was loaded.

This commit make the soved game case and no saved game case consistent.

Defect introduced in b90280855f.

(cherry picked from commit fb9b02a9bac0ac8cad2e473b76d3eda559ee5d66)
Signed-off-by: Itms <itms@wildfiregames.com>