Fix AI errors when promoting an athenian spearman to champion.
Ideally the AI would try to carry on and recover gracefully but the code is setup in such a way that this seems like it would lead to a lot of different bugs down the line.

With help from @langbart on QA and code.

Fixes #7425

(cherry picked from commit c20ca02911579d3c442c538b08c11ade6f8e4f4c)
Signed-off-by: Itms <itms@wildfiregames.com>