Fix crash when trying to join games in lobby dialog
This fixes an oversight from d655455304. During a code deduplication, a
check was removed. When looking at the lobby as a dialog (while in
gamesetup), it should be impossible to try to join a game.

(cherry picked from commit 6ac5c376eab55e59b831605e1b2a01ec03f8dfa8)
Signed-off-by: Itms <itms@wildfiregames.com>