Create port forwarding for correct port with UPnP
Up to know the UPnP logic ignored the port a user was hosting a game on
and always added a port forwarding for the default port UDP 20595. This
commit fixes that, so a port forwarding is added for the actual port a
game is hosted on.

(cherry picked from commit fcd3fc2aa33a45458d080b7ef9a3e02e7d8fa8ba)
Signed-off-by: Itms <itms@wildfiregames.com>