Remove inconsistent override in CDropdown.h
It was added in 670f1e5d42 and, while not illegal, was inconsistent with the
rest of the header file, which caused compiler warnings.
As a quick solution the override keyword is simply removed again.
In the long run, it would be good to still modernise the style of all files
in that directory (at once).

(cherry picked from commit 2e450f0f525e33a22606661d4d596c8567d5d1d8)
Signed-off-by: phosit <phosit@autistici.org>