Close register-page before opening login-page
The `onTick` of the register-page was still running when connecting to
the lobby. This lead to an error when the `connect` message was
received.
With this commit returning from the login-page will always return to the
entrance-page. Previously it returned to the register-page when it was
opened by through that page.

(cherry picked from commit 0b83e240650ba3018830e736d9190974b0cf82c6)
Signed-off-by: phosit <phosit@autistici.org>